Move utils/distribution.py to utils/numeric/distribution
Relocate distribution helpers to inference_perf/utils/numeric/distribution/utils.py
and re-export generate_distribution and sample_from_distribution from the package
__init__.py so callers import from inference_perf.utils.numeric.distribution.
