+++ inference_perf/datagen/shared_prefix_datagen.py
@@ -39,7 +39,7 @@ from inference_perf.payloads import (
     VideoSpecUnion,
 )
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
-from inference_perf.utils.distribution import sample_from_distribution
+from inference_perf.utils.numeric.distribution import sample_from_distribution
 
 from .base import DataGenerator, LazyLoadDataMixin
 from .datagen_utils import (
